GC.isDoomed() and GC.isDestroyed() functions were added
io.murano.system.GC.isDoomed() static method was added. It can be used
within the .destroy method to test if other object is also going to be
destroyed.

io.murano.system.GC.isDestroyed() static method was added. It checks if
the object is destroyed and thus no methods can be invoked on it.

Targets-blueprint: dependency-driven-resource-deallocation
Change-Id: I9ca51f342be27e88e149f217b41145becdfbe232