Use openstack-client variables in openstack CLI calls
In the change https://review.openstack.org/#/c/364885/ was
introduced fix for murano multiregion setup in case of
network determination.
We've explored that this patch may have negative impact
for multiregion installation (like freezes of the installer).

In this patch important variables were changed to openstack CLI
style and completely should fix this problem.

Change-Id: I53533864fdf0f5cc78c2b62b67261974e031fa3e
Related-Bug: #1618999