Prevent executor finalization until exception are not handled
If exception was raised during the deployment early call to
executor.finalize() caused exception stack trace to be garbage
collected. As a result attempt to log it caused another exception.

This commit moves executor finalization to __exit__ method
which is invoked after the logging. In case of success finalize()
is going to be called twice but due to the fact that it clears
the state it doesn't cause any issues

Targets-blueprint: dependency-driven-resource-deallocation
Change-Id: I160dbdfd938ee3eed7ddd27bc0e082a576401fad