Prevent logging of result of resources.string() method call
The result of resources.string() method is the content of the file
saved to string. Logging of binary file content causes
UnicodeDecode error. Moreover, logging of any file content can be
security issue. Finally, it is just not practical to fill the logs
with tons of text from the big files.

Change-Id: I87077b002f2a8888c22b4dfba1f7b9f0508fec69
Closes-bug: #1561522