Increase unit test coverage.
This commit is the first of many that me and my team plan to submit with the goal of increasing murano unit test coverage.
These commits will be submitted in blocks of 500-600 lines to make them more manageable.
This commit focuses on the middleware and v1 components of the API.

Change-Id: I0c727a2e12c6d786b5e74f21c815374034e4cdd1
Implements bp: murano-unit-test-coverage
Co-Authored-By: Felipe Monteiro <felipe.monteiro@att.com>
Co-Authored-By: Samantha Blanco <samantha.blanco@att.com>
Co-Authored-By: Julian Sy <julian.sy@att.com>