Convert =0 version specs to ==0 specs
Murano is using semantic_version library for version specs. 2.3.1
version of the library, which is the minimum allowed for newton does not
treat '=0' as a valid spec, and requires '==0' (double equal signs).
This issue went unnoticed, since there 2.5.0 version of the library
allow both version spec, and it has been only discovered when, when
tested against packaged version of the library.
This commit converts any specs in format =\d.*
into ==.d.* format, thus adding compatibility to 2.3.1 library.

Closes-Bug: #1626238
Change-Id: I68b50726a8b6547cc768452dcfc197e072a8e104