[Docs] Add skeleton for My first Murano app guide
Add skeleton for "My first Murano app" guide for third parties
developers. It is needed for voting job creation in project-config that
builds this guide but does not publish it.

The main purpose of this tutorial is to facilitate the learning process
for Murano App newcome developers and show how it's easy to create a
Murano App for their own cloud applications from scratch. This guide
outlines the creation of Plone Murano App step by step:
https://review.openstack.org/#/c/346492/

Change-Id: Ib184891f7c4c94b0bb2ded92648302eaecad6b5b