Filter enviroment list by project id
Modified a "List Environments" API call to add an ability to filter
environments by an owner project (tenant).
The request now supports extra query parameter - tenant. If this
parameter is passed the list of returned environments is filtered to
return only the environments owned by the tenant with this id.
This operation requires a 'list_environments_all_tenants' policy to be
enabled for a caller user (by default it means that the user has to
have an "admin" role).

Implements-blueprint: list-environments-of-a-given-project
ApiImpact

Change-Id: I90b0dd672d5a431f63e52753d7f517fd3d0ec6eb