Increase unit test coverage for Stats API.
Please note the comment in test_update.
The functionality this test tests for appears
to be incorrect. That is, it looks as if
test_update should update the host for the
Statistics object that is passed in, but
does not. Instead, it simply saves the
Statistics object that is passed in.
Thus, the function being tested appears
to contradict its intended use case.

Implements bp: murano-unit-test-coverage

Change-Id: Ie6a07ef5c0f345dbb4bbb93defb36b2549697a97