Improve first app builds
The complete html tree gets published but the .doctree is not needed for
publishing, place it outside of the html tree so that it won't pushed to
the website.

Also, only build html documents.

Fix bug_project, the report a bug URL should point to murano and not to
the api project.

Change-Id: Icf0c901257b452d216d293eb3f1c7b447b58201d