Configuration is now properly applied to new nodes
If the configuration of software component has not been changed, but a
set of nodes in the server group has been modified, the component's
default checkClusterIsConfigured method will now properly return true,
thus the configuration will be applied on newly added nodes.

This is achieved by storing instance ids as part of component's
'configuration' attributed stored for its server group.

Change-Id: Ic8bbddc577518071d90a6e33518156047a1d2e2e
Closes-bug: #1634206