Fixed SharedIp class
Since the introduction of multi-region capabilities the
`io.murano.SharedIp` class was not working, since it utilized but
didn't define the 'getRegion' method, and directly used the
defaultNetworks.environment property of the Environment assuming it is
a class (yet it became a template quite recently).

Both these issues have been addressed so the shared ips may be used
again.

Change-Id: I50075b03cb2b61d97bf23300478c07d9d5d774cf
Closes-bug: #1645724