Fixes TypeError thrown by parallel_select in DSL Helpers.
This applies to Python 3.0+ environments only. When an exception
is re-reraised using six, an instance of the exception is
passed as the first argument, which causes six.reraise to
thrown a TypeError exception, after trying to instantiate
an exception from an instance of an exception, rather than
from an exception class.

This fix simply changes the first argument to an exception
class, instead of exception instance.

Change-Id: I76f1ab6f1540880e9b8596071244b76f4f7478a0
Closes-Bug: #1651795
Related-Bug: #1638722