Update rename environment API endpoint
The ``update`` method in the environments API is only responsible
for renaming the environment.

However, the entire request body is passed to the ``update``
DB method if the name of the environment meets the initial checks
(enough characters and not empty string).

While this is not a problem if the request body is generated
through the CLI client or the dashboard, it is a problem using
something like curl, in which other attributes in the environment
could be added to the request body, causing the environments API
to violate its contract: an endpoint that is supposed to rename
an environment should only ever rename the environment.

Change-Id: If6314f704c22ff4d61411b1808d6d73bfb7ac562