Ensure assigned-ips output is present in Heat template with Nova Network
When using Murano with Nova Network, the assigned-ips output is only added to
the template when a floating IP is assigned. This makes the deployment fail
when the engine tries to get the list of IPs of the created instance.

This patch makes sure that the hostname-assigned-ips output is added to the
template also when a floating IP is not assigned.

Change-Id: I8970a34e7c0b9fefdc236b046d6506f778143f3a
Closes-Bug: 1710623
(cherry picked from commit 3990acfcce6b9f1dd9bde9ceb7e0817c19acff34)