Remove separate firstapp build
With the large reorg of the docs site, this can be easily *only* as
part of the murano documentation, there's no benefit for publishing an
extra document elsewhere. Fully integrate the documents into the manual.

Note that the separate build target was not linked from any of our index
sites.

Remove extra title from README.rst, it let to duplicated title in main
table of contents.

Remove also post job for publishing the firstapp document - it currently fails -
and is now not needed anymore.

Change-Id: Ib5259907b3f999062042e483ac81e6aed17705ca