Remove murano-tempest-plugin entry from setup.cfg
As murano-tempest-plugin already been moved to separate
repo, it has to be removed from murano setup.cfg.

Due to this it will force murano installer to install murano-tempest-plugin
as Temepst will discover this plugin and complain about its missing package.

Congress is getting this error and their gate is blocked due to this.
http://logs.openstack.org/57/508657/22/check/congress-devstack-api-mysql/1ad845d/job-output.txt.gz#_2017-12-07_19_33_07_219034

Change-Id: Iab3a851fdb9001ad39ef443be12701e0e2a75e5a
Closes-Bug: #1737123