Remove python-dev from list of preinstalled packages
python-dev package causes cloud-init to fail on Fedora-based
distros because the package is called python-devel there.
Since this package is present in python-pip recommended dependencies
and thus going to be installed anyway, there is no need to include
it explicitly. With this fix it becomes possible to use bare
Fedora cloud image with murano.

Change-Id: I276839b9ba772d8f0a6bac975b204b0308839170