Fix Contract on Project property, extra.
The current contract for this property (in Project.yaml) is:

      extra:
        Contract:
          $.string().notNull(): $.string()

Deploying my Murano application, the code attempts to populate this
property with:

     u'tags': []

and gets this error:

    ContractViolationException: [io.murano.Project.extra['tags']] Value [] violates string() contract

My fix is:

        extra:
        Contract:
          $.string().notNull(): $

This allows the 2nd value of the pair to be any scalar or data structure.  A string
or a list of strings is acceptable now.

Change-Id: I1b0438ad451db2b7a89a9fb65dc948b52ceb2f3f
Co-Authored-By: Devjani Ray <devjani.ray@oracle.com>
Closes-Bug: 1762821