Add local bindep.txt
As it was announced [1] global bindep-fallback.txt was removed and now
projects need to have a local bindep.txt to be able to install binary
dependencies for testing.

In test jobs the script tools/test-setup.sh is called which requires
mysql and postgres servers and clients to be installed.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-June/007272.html

Change-Id: I37d64aca485eb81ac16d7d85d5b695e573c8793c
(cherry picked from commit ddd5378ac8535b48e5ae2885b3187a5b1b88b766)